fix(review comment: stabilize filter URLs by serializing entity names instead of documentIds
